requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0

drf-spectacular>=0.28.0
django-cors-headers>=4.9.0

vosk>=0.3.44
faster_whisper>=1.2
whitenoise>=6.7
python-dotenv>=1.1.1
alt_profanity_check>=1

pandas
